All Field Names in a clinicaltrials.gov Record

Character vector containing all queriable field names in a clinicaltrials.gov record

Usage

data(all_fields)

Format

A character vector with the names of all 322 queriable fields

Source

https://clinicaltrials.gov/api/info/study_fields_list?fmt=JSON

Examples

all_fields
#>   [1] "Acronym"                                
#>   [2] "AgreementOtherDetails"                  
#>   [3] "AgreementPISponsorEmployee"             
#>   [4] "AgreementRestrictionType"               
#>   [5] "AgreementRestrictiveAgreement"          
#>   [6] "ArmGroupDescription"                    
#>   [7] "ArmGroupInterventionName"               
#>   [8] "ArmGroupLabel"                          
#>   [9] "ArmGroupType"                           
#>  [10] "AvailIPDComment"                        
#>  [11] "AvailIPDId"                             
#>  [12] "AvailIPDType"                           
#>  [13] "AvailIPDURL"                            
#>  [14] "BaselineCategoryTitle"                  
#>  [15] "BaselineClassDenomCountGroupId"         
#>  [16] "BaselineClassDenomCountValue"           
#>  [17] "BaselineClassDenomUnits"                
#>  [18] "BaselineClassTitle"                     
#>  [19] "BaselineDenomCountGroupId"              
#>  [20] "BaselineDenomCountValue"                
#>  [21] "BaselineDenomUnits"                     
#>  [22] "BaselineGroupDescription"               
#>  [23] "BaselineGroupId"                        
#>  [24] "BaselineGroupTitle"                     
#>  [25] "BaselineMeasureCalculatePct"            
#>  [26] "BaselineMeasureDenomCountGroupId"       
#>  [27] "BaselineMeasureDenomCountValue"         
#>  [28] "BaselineMeasureDenomUnits"              
#>  [29] "BaselineMeasureDenomUnitsSelected"      
#>  [30] "BaselineMeasureDescription"             
#>  [31] "BaselineMeasureDispersionType"          
#>  [32] "BaselineMeasureParamType"               
#>  [33] "BaselineMeasurePopulationDescription"   
#>  [34] "BaselineMeasureTitle"                   
#>  [35] "BaselineMeasureUnitOfMeasure"           
#>  [36] "BaselineMeasurementComment"             
#>  [37] "BaselineMeasurementGroupId"             
#>  [38] "BaselineMeasurementLowerLimit"          
#>  [39] "BaselineMeasurementSpread"              
#>  [40] "BaselineMeasurementUpperLimit"          
#>  [41] "BaselineMeasurementValue"               
#>  [42] "BaselinePopulationDescription"          
#>  [43] "BaselineTypeUnitsAnalyzed"              
#>  [44] "BioSpecDescription"                     
#>  [45] "BioSpecRetention"                       
#>  [46] "BriefSummary"                           
#>  [47] "BriefTitle"                             
#>  [48] "CentralContactEMail"                    
#>  [49] "CentralContactName"                     
#>  [50] "CentralContactPhone"                    
#>  [51] "CentralContactPhoneExt"                 
#>  [52] "CentralContactRole"                     
#>  [53] "CollaboratorClass"                      
#>  [54] "CollaboratorName"                       
#>  [55] "CompletionDate"                         
#>  [56] "CompletionDateType"                     
#>  [57] "Condition"                              
#>  [58] "ConditionAncestorId"                    
#>  [59] "ConditionAncestorTerm"                  
#>  [60] "ConditionBrowseBranchAbbrev"            
#>  [61] "ConditionBrowseBranchName"              
#>  [62] "ConditionBrowseLeafAsFound"             
#>  [63] "ConditionBrowseLeafId"                  
#>  [64] "ConditionBrowseLeafName"                
#>  [65] "ConditionBrowseLeafRelevance"           
#>  [66] "ConditionMeshId"                        
#>  [67] "ConditionMeshTerm"                      
#>  [68] "DelayedPosting"                         
#>  [69] "DesignAllocation"                       
#>  [70] "DesignInterventionModel"                
#>  [71] "DesignInterventionModelDescription"     
#>  [72] "DesignMasking"                          
#>  [73] "DesignMaskingDescription"               
#>  [74] "DesignObservationalModel"               
#>  [75] "DesignPrimaryPurpose"                   
#>  [76] "DesignTimePerspective"                  
#>  [77] "DesignWhoMasked"                        
#>  [78] "DetailedDescription"                    
#>  [79] "DispFirstPostDate"                      
#>  [80] "DispFirstPostDateType"                  
#>  [81] "DispFirstSubmitDate"                    
#>  [82] "DispFirstSubmitQCDate"                  
#>  [83] "EligibilityCriteria"                    
#>  [84] "EnrollmentCount"                        
#>  [85] "EnrollmentType"                         
#>  [86] "EventGroupDeathsNumAffected"            
#>  [87] "EventGroupDeathsNumAtRisk"              
#>  [88] "EventGroupDescription"                  
#>  [89] "EventGroupId"                           
#>  [90] "EventGroupOtherNumAffected"             
#>  [91] "EventGroupOtherNumAtRisk"               
#>  [92] "EventGroupSeriousNumAffected"           
#>  [93] "EventGroupSeriousNumAtRisk"             
#>  [94] "EventGroupTitle"                        
#>  [95] "EventsDescription"                      
#>  [96] "EventsFrequencyThreshold"               
#>  [97] "EventsTimeFrame"                        
#>  [98] "ExpAccTypeIndividual"                   
#>  [99] "ExpAccTypeIntermediate"                 
#> [100] "ExpAccTypeTreatment"                    
#> [101] "ExpandedAccessNCTId"                    
#> [102] "ExpandedAccessStatusForNCTId"           
#> [103] "FDAAA801Violation"                      
#> [104] "FlowAchievementComment"                 
#> [105] "FlowAchievementGroupId"                 
#> [106] "FlowAchievementNumSubjects"             
#> [107] "FlowAchievementNumUnits"                
#> [108] "FlowDropWithdrawComment"                
#> [109] "FlowDropWithdrawType"                   
#> [110] "FlowGroupDescription"                   
#> [111] "FlowGroupId"                            
#> [112] "FlowGroupTitle"                         
#> [113] "FlowMilestoneComment"                   
#> [114] "FlowMilestoneType"                      
#> [115] "FlowPeriodTitle"                        
#> [116] "FlowPreAssignmentDetails"               
#> [117] "FlowReasonComment"                      
#> [118] "FlowReasonGroupId"                      
#> [119] "FlowReasonNumSubjects"                  
#> [120] "FlowReasonNumUnits"                     
#> [121] "FlowRecruitmentDetails"                 
#> [122] "FlowTypeUnitsAnalyzed"                  
#> [123] "Gender"                                 
#> [124] "GenderBased"                            
#> [125] "GenderDescription"                      
#> [126] "HasExpandedAccess"                      
#> [127] "HealthyVolunteers"                      
#> [128] "IPDSharing"                             
#> [129] "IPDSharingAccessCriteria"               
#> [130] "IPDSharingDescription"                  
#> [131] "IPDSharingInfoType"                     
#> [132] "IPDSharingTimeFrame"                    
#> [133] "IPDSharingURL"                          
#> [134] "InterventionAncestorId"                 
#> [135] "InterventionAncestorTerm"               
#> [136] "InterventionArmGroupLabel"              
#> [137] "InterventionBrowseBranchAbbrev"         
#> [138] "InterventionBrowseBranchName"           
#> [139] "InterventionBrowseLeafAsFound"          
#> [140] "InterventionBrowseLeafId"               
#> [141] "InterventionBrowseLeafName"             
#> [142] "InterventionBrowseLeafRelevance"        
#> [143] "InterventionDescription"                
#> [144] "InterventionMeshId"                     
#> [145] "InterventionMeshTerm"                   
#> [146] "InterventionName"                       
#> [147] "InterventionOtherName"                  
#> [148] "InterventionType"                       
#> [149] "IsFDARegulatedDevice"                   
#> [150] "IsFDARegulatedDrug"                     
#> [151] "IsPPSD"                                 
#> [152] "IsUSExport"                             
#> [153] "IsUnapprovedDevice"                     
#> [154] "Keyword"                                
#> [155] "LargeDocDate"                           
#> [156] "LargeDocFilename"                       
#> [157] "LargeDocHasICF"                         
#> [158] "LargeDocHasProtocol"                    
#> [159] "LargeDocHasSAP"                         
#> [160] "LargeDocLabel"                          
#> [161] "LargeDocTypeAbbrev"                     
#> [162] "LargeDocUploadDate"                     
#> [163] "LastKnownStatus"                        
#> [164] "LastUpdatePostDate"                     
#> [165] "LastUpdatePostDateType"                 
#> [166] "LastUpdateSubmitDate"                   
#> [167] "LeadSponsorClass"                       
#> [168] "LeadSponsorName"                        
#> [169] "LimitationsAndCaveatsDescription"       
#> [170] "LocationCity"                           
#> [171] "LocationContactEMail"                   
#> [172] "LocationContactName"                    
#> [173] "LocationContactPhone"                   
#> [174] "LocationContactPhoneExt"                
#> [175] "LocationContactRole"                    
#> [176] "LocationCountry"                        
#> [177] "LocationFacility"                       
#> [178] "LocationState"                          
#> [179] "LocationStatus"                         
#> [180] "LocationZip"                            
#> [181] "MaximumAge"                             
#> [182] "MinimumAge"                             
#> [183] "NCTId"                                  
#> [184] "NCTIdAlias"                             
#> [185] "OfficialTitle"                          
#> [186] "OrgClass"                               
#> [187] "OrgFullName"                            
#> [188] "OrgStudyId"                             
#> [189] "OrgStudyIdDomain"                       
#> [190] "OrgStudyIdLink"                         
#> [191] "OrgStudyIdType"                         
#> [192] "OtherEventAssessmentType"               
#> [193] "OtherEventNotes"                        
#> [194] "OtherEventOrganSystem"                  
#> [195] "OtherEventSourceVocabulary"             
#> [196] "OtherEventStatsGroupId"                 
#> [197] "OtherEventStatsNumAffected"             
#> [198] "OtherEventStatsNumAtRisk"               
#> [199] "OtherEventStatsNumEvents"               
#> [200] "OtherEventTerm"                         
#> [201] "OtherOutcomeDescription"                
#> [202] "OtherOutcomeMeasure"                    
#> [203] "OtherOutcomeTimeFrame"                  
#> [204] "OutcomeAnalysisCILowerLimit"            
#> [205] "OutcomeAnalysisCILowerLimitComment"     
#> [206] "OutcomeAnalysisCINumSides"              
#> [207] "OutcomeAnalysisCIPctValue"              
#> [208] "OutcomeAnalysisCIUpperLimit"            
#> [209] "OutcomeAnalysisCIUpperLimitComment"     
#> [210] "OutcomeAnalysisDispersionType"          
#> [211] "OutcomeAnalysisDispersionValue"         
#> [212] "OutcomeAnalysisEstimateComment"         
#> [213] "OutcomeAnalysisGroupDescription"        
#> [214] "OutcomeAnalysisGroupId"                 
#> [215] "OutcomeAnalysisNonInferiorityComment"   
#> [216] "OutcomeAnalysisNonInferiorityType"      
#> [217] "OutcomeAnalysisOtherAnalysisDescription"
#> [218] "OutcomeAnalysisPValue"                  
#> [219] "OutcomeAnalysisPValueComment"           
#> [220] "OutcomeAnalysisParamType"               
#> [221] "OutcomeAnalysisParamValue"              
#> [222] "OutcomeAnalysisStatisticalComment"      
#> [223] "OutcomeAnalysisStatisticalMethod"       
#> [224] "OutcomeAnalysisTestedNonInferiority"    
#> [225] "OutcomeCategoryTitle"                   
#> [226] "OutcomeClassDenomCountGroupId"          
#> [227] "OutcomeClassDenomCountValue"            
#> [228] "OutcomeClassDenomUnits"                 
#> [229] "OutcomeClassTitle"                      
#> [230] "OutcomeDenomCountGroupId"               
#> [231] "OutcomeDenomCountValue"                 
#> [232] "OutcomeDenomUnits"                      
#> [233] "OutcomeGroupDescription"                
#> [234] "OutcomeGroupId"                         
#> [235] "OutcomeGroupTitle"                      
#> [236] "OutcomeMeasureAnticipatedPostingDate"   
#> [237] "OutcomeMeasureCalculatePct"             
#> [238] "OutcomeMeasureDenomUnitsSelected"       
#> [239] "OutcomeMeasureDescription"              
#> [240] "OutcomeMeasureDispersionType"           
#> [241] "OutcomeMeasureParamType"                
#> [242] "OutcomeMeasurePopulationDescription"    
#> [243] "OutcomeMeasureReportingStatus"          
#> [244] "OutcomeMeasureTimeFrame"                
#> [245] "OutcomeMeasureTitle"                    
#> [246] "OutcomeMeasureType"                     
#> [247] "OutcomeMeasureTypeUnitsAnalyzed"        
#> [248] "OutcomeMeasureUnitOfMeasure"            
#> [249] "OutcomeMeasurementComment"              
#> [250] "OutcomeMeasurementGroupId"              
#> [251] "OutcomeMeasurementLowerLimit"           
#> [252] "OutcomeMeasurementSpread"               
#> [253] "OutcomeMeasurementUpperLimit"           
#> [254] "OutcomeMeasurementValue"                
#> [255] "OverallOfficialAffiliation"             
#> [256] "OverallOfficialName"                    
#> [257] "OverallOfficialRole"                    
#> [258] "OverallStatus"                          
#> [259] "OversightHasDMC"                        
#> [260] "PatientRegistry"                        
#> [261] "Phase"                                  
#> [262] "PointOfContactEMail"                    
#> [263] "PointOfContactOrganization"             
#> [264] "PointOfContactPhone"                    
#> [265] "PointOfContactPhoneExt"                 
#> [266] "PointOfContactTitle"                    
#> [267] "PrimaryCompletionDate"                  
#> [268] "PrimaryCompletionDateType"              
#> [269] "PrimaryOutcomeDescription"              
#> [270] "PrimaryOutcomeMeasure"                  
#> [271] "PrimaryOutcomeTimeFrame"                
#> [272] "ReferenceCitation"                      
#> [273] "ReferencePMID"                          
#> [274] "ReferenceType"                          
#> [275] "RemovedCountry"                         
#> [276] "ResponsiblePartyInvestigatorAffiliation"
#> [277] "ResponsiblePartyInvestigatorFullName"   
#> [278] "ResponsiblePartyInvestigatorTitle"      
#> [279] "ResponsiblePartyOldNameTitle"           
#> [280] "ResponsiblePartyOldOrganization"        
#> [281] "ResponsiblePartyType"                   
#> [282] "ResultsFirstPostDate"                   
#> [283] "ResultsFirstPostDateType"               
#> [284] "ResultsFirstSubmitDate"                 
#> [285] "ResultsFirstSubmitQCDate"               
#> [286] "RetractionPMID"                         
#> [287] "RetractionSource"                       
#> [288] "SamplingMethod"                         
#> [289] "SecondaryId"                            
#> [290] "SecondaryIdDomain"                      
#> [291] "SecondaryIdLink"                        
#> [292] "SecondaryIdType"                        
#> [293] "SecondaryOutcomeDescription"            
#> [294] "SecondaryOutcomeMeasure"                
#> [295] "SecondaryOutcomeTimeFrame"              
#> [296] "SeeAlsoLinkLabel"                       
#> [297] "SeeAlsoLinkURL"                         
#> [298] "SeriousEventAssessmentType"             
#> [299] "SeriousEventNotes"                      
#> [300] "SeriousEventOrganSystem"                
#> [301] "SeriousEventSourceVocabulary"           
#> [302] "SeriousEventStatsGroupId"               
#> [303] "SeriousEventStatsNumAffected"           
#> [304] "SeriousEventStatsNumAtRisk"             
#> [305] "SeriousEventStatsNumEvents"             
#> [306] "SeriousEventTerm"                       
#> [307] "StartDate"                              
#> [308] "StartDateType"                          
#> [309] "StatusVerifiedDate"                     
#> [310] "StdAge"                                 
#> [311] "StudyFirstPostDate"                     
#> [312] "StudyFirstPostDateType"                 
#> [313] "StudyFirstSubmitDate"                   
#> [314] "StudyFirstSubmitQCDate"                 
#> [315] "StudyPopulation"                        
#> [316] "StudyType"                              
#> [317] "TargetDuration"                         
#> [318] "UnpostedEventDate"                      
#> [319] "UnpostedEventType"                      
#> [320] "UnpostedResponsibleParty"               
#> [321] "VersionHolder"                          
#> [322] "WhyStopped"