Registration Data Elements Fields

Character vector containing registration data elements fields, these elements are all uploaded before the study is completed

Usage

data(registration_fields)

Format

A character vector

Examples

registration_fields
#>  [1] "NCTId"                               
#>  [2] "OrgStudyId"                          
#>  [3] "BriefTitle"                          
#>  [4] "Acronym"                             
#>  [5] "OfficialTitle"                       
#>  [6] "StudyType"                           
#>  [7] "OverallStatus"                       
#>  [8] "WhyStopped"                          
#>  [9] "StartDate"                           
#> [10] "CompletionDate"                      
#> [11] "ResponsiblePartyType"                
#> [12] "ResponsiblePartyInvestigatorFullName"
#> [13] "LeadSponsorName"                     
#> [14] "IsFDARegulatedDrug"                  
#> [15] "IsFDARegulatedDevice"                
#> [16] "IsUnapprovedDevice"                  
#> [17] "OversightHasDMC"                     
#> [18] "BriefSummary"                        
#> [19] "DetailedDescription"                 
#> [20] "Condition"                           
#> [21] "Keyword"                             
#> [22] "DesignPrimaryPurpose"                
#> [23] "Phase"                               
#> [24] "DesignInterventionModel"             
#> [25] "DesignInterventionModelDescription"  
#> [26] "DesignWhoMasked"                     
#> [27] "DesignMasking"                       
#> [28] "DesignMaskingDescription"            
#> [29] "DesignAllocation"                    
#> [30] "EnrollmentCount"                     
#> [31] "DesignObservationalModel"            
#> [32] "DesignTimePerspective"               
#> [33] "BioSpecRetention"                    
#> [34] "BioSpecDescription"                  
#> [35] "TargetDuration"                      
#> [36] "ArmGroupLabel"                       
#> [37] "ArmGroupType"                        
#> [38] "ArmGroupDescription"                 
#> [39] "InterventionType"                    
#> [40] "InterventionName"                    
#> [41] "InterventionDescription"             
#> [42] "PrimaryOutcomeMeasure"               
#> [43] "PrimaryOutcomeDescription"           
#> [44] "PrimaryOutcomeTimeFrame"             
#> [45] "SecondaryOutcomeMeasure"             
#> [46] "SecondaryOutcomeDescription"         
#> [47] "SecondaryOutcomeTimeFrame"           
#> [48] "OtherOutcomeMeasure"                 
#> [49] "OtherOutcomeDescription"             
#> [50] "OtherOutcomeTimeFrame"               
#> [51] "Gender"                              
#> [52] "GenderBased"                         
#> [53] "GenderDescription"                   
#> [54] "MinimumAge"                          
#> [55] "MaximumAge"                          
#> [56] "HealthyVolunteers"                   
#> [57] "EligibilityCriteria"                 
#> [58] "StudyPopulation"                     
#> [59] "SamplingMethod"